+++ app/src/main/java/za/kilowatch/ultimatefilemanager/settings/IconPackExportActivity.kt
@@ -124,7 +124,7 @@ class IconPackExportActivity : AppCompatActivity() {
             listOf("toolbar_create_new", "toolbar_copy", "toolbar_move", "toolbar_rename", "toolbar_share",
                 "toolbar_copy_encrypt", "toolbar_move_encrypt", "toolbar_favorite",
                 "toolbar_hide", "toolbar_unhide", "toolbar_select_all", "toolbar_compress",
-                "toolbar_image_compress", "toolbar_delete")
+                "toolbar_image_compress", "toolbar_delete", "toolbar_pin", "toolbar_unpin")
         ))
 
         // Navigation